// src/common/enums/error-code.enum.ts
export enum ErrorCode {
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  VALIDATION_ERROR = 'VALIDATION_ERROR',
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN',
  NOT_FOUND = 'NOT_FOUND',
  CONFLICT = 'CONFLICT',
  BAD_REQUEST = 'BAD_REQUEST',
  AUTH_INVALID_CREDENTIALS = 'AUTH_INVALID_CREDENTIALS',
  AUTH_EMAIL_ALREADY_EXISTS = 'AUTH_EMAIL_ALREADY_EXISTS',
  AUTH_TOKEN_EXPIRED = 'AUTH_TOKEN_EXPIRED',
  AUTH_TOKEN_INVALID = 'AUTH_TOKEN_INVALID',
  AUTH_USER_NOT_FOUND = 'AUTH_USER_NOT_FOUND',
  CLIENT_NOT_FOUND = 'CLIENT_NOT_FOUND',
  CLIENT_EMAIL_ALREADY_EXISTS = 'CLIENT_EMAIL_ALREADY_EXISTS',
  CLIENT_SUMMARY_MIXED_CURRENCY = 'CLIENT_SUMMARY_MIXED_CURRENCY',
  AGREEMENT_NOT_FOUND = 'AGREEMENT_NOT_FOUND',
  AGREEMENT_NOT_DRAFT = 'AGREEMENT_NOT_DRAFT',
  AGREEMENT_ALREADY_SENT = 'AGREEMENT_ALREADY_SENT',
  AGREEMENT_ALREADY_APPROVED = 'AGREEMENT_ALREADY_APPROVED',
  AGREEMENT_CANNOT_BE_MODIFIED = 'AGREEMENT_CANNOT_BE_MODIFIED',
  AGREEMENT_CLIENT_REQUIRED = 'AGREEMENT_CLIENT_REQUIRED',
  AGREEMENT_POLICY_REQUIRED = 'AGREEMENT_POLICY_REQUIRED',
  POLICY_NOT_FOUND = 'POLICY_NOT_FOUND',
  POLICY_INVALID_REVIEW_PERIOD = 'POLICY_INVALID_REVIEW_PERIOD',
  POLICY_INVALID_CONTENT = 'POLICY_INVALID_CONTENT',
  MILESTONE_NOT_FOUND = 'MILESTONE_NOT_FOUND',
  MILESTONE_INVALID_AMOUNT = 'MILESTONE_INVALID_AMOUNT',
  MILESTONE_INVALID_ORDER = 'MILESTONE_INVALID_ORDER',
  MILESTONE_ALREADY_SUBMITTED = 'MILESTONE_ALREADY_SUBMITTED',
  MILESTONE_NOT_ACTIVE = 'MILESTONE_NOT_ACTIVE',
  MILESTONE_PAYMENT_NOT_RESERVED = 'MILESTONE_PAYMENT_NOT_RESERVED',
  MILESTONE_PAYMENT_NOT_WAITING = 'MILESTONE_PAYMENT_NOT_WAITING',
  PAYMENT_NOT_FOUND = 'PAYMENT_NOT_FOUND',
  PAYMENT_ALREADY_RESERVED = 'PAYMENT_ALREADY_RESERVED',
  PAYMENT_ALREADY_RELEASED = 'PAYMENT_ALREADY_RELEASED',
  PAYMENT_NOT_READY_TO_RELEASE = 'PAYMENT_NOT_READY_TO_RELEASE',
  PAYMENT_INVALID_AMOUNT = 'PAYMENT_INVALID_AMOUNT',
  PAYMENT_DEMO_MODE_ONLY = 'PAYMENT_DEMO_MODE_ONLY',
  PAYMENT_INVALID_TRANSITION = 'PAYMENT_INVALID_TRANSITION',
  PAYMENT_MILESTONE_REQUIRED = 'PAYMENT_MILESTONE_REQUIRED',
  DELIVERY_NOT_FOUND = 'DELIVERY_NOT_FOUND',
  DELIVERY_ALREADY_SUBMITTED = 'DELIVERY_ALREADY_SUBMITTED',
  DELIVERY_URL_OR_FILE_REQUIRED = 'DELIVERY_URL_OR_FILE_REQUIRED',
  DELIVERY_NOT_IN_REVIEW = 'DELIVERY_NOT_IN_REVIEW',
  DELIVERY_ALREADY_ACCEPTED = 'DELIVERY_ALREADY_ACCEPTED',
  DELIVERY_ALREADY_EXISTS = 'DELIVERY_ALREADY_EXISTS',
  DELIVERY_NOT_EDITABLE = 'DELIVERY_NOT_EDITABLE',
  DELIVERY_NOT_SUBMITTABLE = 'DELIVERY_NOT_SUBMITTABLE',
  DELIVERY_EVIDENCE_REQUIRED = 'DELIVERY_EVIDENCE_REQUIRED',
  DELIVERY_NOT_REVIEWABLE = 'DELIVERY_NOT_REVIEWABLE',
  AGREEMENT_NOT_ACTIVE = 'AGREEMENT_NOT_ACTIVE',
  PAYMENT_NOT_RESERVED = 'PAYMENT_NOT_RESERVED',
  PORTAL_TOKEN_INVALID = 'PORTAL_TOKEN_INVALID',
  PORTAL_TOKEN_EXPIRED = 'PORTAL_TOKEN_EXPIRED',
  PORTAL_TOKEN_REVOKED = 'PORTAL_TOKEN_REVOKED',
  PORTAL_AGREEMENT_NOT_FOUND = 'PORTAL_AGREEMENT_NOT_FOUND',
  PORTAL_ACTION_NOT_ALLOWED = 'PORTAL_ACTION_NOT_ALLOWED',
  AI_PLAN_GENERATION_FAILED = 'AI_PLAN_GENERATION_FAILED',
  AI_REVIEW_FAILED = 'AI_REVIEW_FAILED',
  AI_REVIEW_NOT_FOUND = 'AI_REVIEW_NOT_FOUND',
  AI_REVIEW_ALREADY_COMPLETED = 'AI_REVIEW_ALREADY_COMPLETED',
  AI_INVALID_RESPONSE = 'AI_INVALID_RESPONSE',
  CHANGE_REQUEST_NOT_FOUND = 'CHANGE_REQUEST_NOT_FOUND',
  CHANGE_REQUEST_AMOUNT_INVALID = 'CHANGE_REQUEST_AMOUNT_INVALID',
  CHANGE_REQUEST_NOT_EDITABLE = 'CHANGE_REQUEST_NOT_EDITABLE',
  CHANGE_REQUEST_NOT_SENDABLE = 'CHANGE_REQUEST_NOT_SENDABLE',
  CHANGE_REQUEST_NOT_APPROVABLE = 'CHANGE_REQUEST_NOT_APPROVABLE',
  CHANGE_REQUEST_NOT_DECLINABLE = 'CHANGE_REQUEST_NOT_DECLINABLE',
  CHANGE_REQUEST_NOT_APPROVED = 'CHANGE_REQUEST_NOT_APPROVED',
  PAYMENT_NOT_FUNDABLE = 'PAYMENT_NOT_FUNDABLE',
  AI_REVIEW_NOT_ELIGIBLE_FOR_CHANGE_REQUEST = 'AI_REVIEW_NOT_ELIGIBLE_FOR_CHANGE_REQUEST',
  EMAIL_SEND_FAILED = 'EMAIL_SEND_FAILED',
  EMAIL_TEMPLATE_NOT_FOUND = 'EMAIL_TEMPLATE_NOT_FOUND',
  EMAIL_RENDER_FAILED = 'EMAIL_RENDER_FAILED',
  EMAIL_RECIPIENT_REQUIRED = 'EMAIL_RECIPIENT_REQUIRED',
  CLIENT_EMAIL_MISSING = 'CLIENT_EMAIL_MISSING',
  EMAIL_TYPE_NOT_SUPPORTED = 'EMAIL_TYPE_NOT_SUPPORTED',
  EMAIL_CONTEXT_INCOMPLETE = 'EMAIL_CONTEXT_INCOMPLETE',
  EMAIL_NOTIFICATIONS_DISABLED = 'EMAIL_NOTIFICATIONS_DISABLED',
  AGREEMENT_NOT_INVITABLE = 'AGREEMENT_NOT_INVITABLE',
  SETTINGS_NOT_FOUND = 'SETTINGS_NOT_FOUND',
  SETTINGS_INVALID_VALUE = 'SETTINGS_INVALID_VALUE',
  DASHBOARD_RANGE_INVALID = 'DASHBOARD_RANGE_INVALID',
  DASHBOARD_AGGREGATION_FAILED = 'DASHBOARD_AGGREGATION_FAILED',
  TIMELINE_EVENT_TYPE_INVALID = 'TIMELINE_EVENT_TYPE_INVALID',
  TIMELINE_METADATA_INVALID = 'TIMELINE_METADATA_INVALID',
  PORTAL_TOKEN_CREATE_FAILED = 'PORTAL_TOKEN_CREATE_FAILED',
  AGREEMENT_NOT_APPROVABLE = 'AGREEMENT_NOT_APPROVABLE',
  AGREEMENT_NOT_CHANGEABLE = 'AGREEMENT_NOT_CHANGEABLE',
  AGREEMENT_NOT_REJECTABLE = 'AGREEMENT_NOT_REJECTABLE',
}